🔨Refactor: GNB
